ref: Refactor ExternalIssueList to make it easier to re-skin for User Feedback
